refine tests and extra dependencies
