+++ src/promptflow/tests/sdk_cli_azure_test/e2etests/test_flow_serve.py
@@ -82,4 +82,17 @@ def test_azureml_serving_api_with_connection_name_override(serving_client_with_c
     assert (
         response.status_code == 200
     ), f"Response code indicates error {response.status_code} - {response.data.decode()}"
-    assert "output_prompt" in json.loads(response.data.decode())
+    assert "api_base" not in json.loads(response.data.decode()).values()
+
+
+@pytest.mark.usefixtures("serving_client_with_connection_data_override")
+@pytest.mark.e2etest
+def test_azureml_serving_api_with_connection_data_override(serving_client_with_connection_data_override):
+    response = serving_client_with_connection_data_override.get("/health")
+    assert b'{"status":"Healthy","version":"0.0.1"}' in response.data
+
+    response = serving_client_with_connection_data_override.post("/score", data=json.dumps({"text": "hi"}))
+    assert (
+        response.status_code == 200
+    ), f"Response code indicates error {response.status_code} - {response.data.decode()}"
+    assert "api_base" in json.loads(response.data.decode()).values()
