+++ src/promptflow/tests/test_configs/flows/llm_connection_override/conn_tool.py
@@ -1,0 +1,8 @@
+from promptflow import tool
+from promptflow.connections import AzureOpenAIConnection
+
+
+@tool
+def conn_tool(conn: AzureOpenAIConnection):
+    assert isinstance(conn, AzureOpenAIConnection)
+    return conn.api_base
