--- src/promptflow/tests/test_configs/flows/llm_connection_override/inputs.json
@@ -1,3 +1,0 @@
-{
-    "text": "The event sourcing pattern involves using an append-only store to record the full series of actions on that data. The Azure Cosmos DB change feed is a great choice as a central data store in event sourcing architectures in which all data ingestion is modeled as writes (no updates or deletes). In this case, each write to Azure Cosmos DB is an \"event,\" so there's a full record of past events in the change feed. Typical uses of the events published by the central event store are to maintain materialized views or to integrate with external systems. Because there's no time limit for retention in the change feed latest version mode, you can replay all past events by reading from the beginning of your Azure Cosmos DB container's change feed. You can even have multiple change feed consumers subscribe to the same container's change feed."
-}
