--- src/promptflow/tests/test_configs/flows/llm_connection_override/summarize_text_content_prompt.jinja2
@@ -1,7 +1,0 @@
-Please summarize the following text in one paragraph. 50 words.
-
-Do not add any information that is not in the text.
-
-Text: {{text}}
-
-Summary:
